+++ doc/whats_new.rst
@@ -25,6 +25,8 @@ CLI
 - Deprecate the ``--slurm`` parameter which will be removed in benchopt 1.8.
   By `Thomas Moreau`_ (:gh:`673`).
 
+- Improved output formatting for benchmark ``run/install/test``.
+  By `Thomas Moreau`_ (:gh:`847`).
 
 API
 ---
@@ -32,8 +34,11 @@ API
 FIX
 ---
 
-- Fix parsing of the parameters when a class is not installed.
-  By `Thomas Moreau`_ (:gh:`844`)
+- Improve AST parsing when missing attributes in the class.
+  By `Thomas Moreau`_ (:gh:`846`)
+
+- Fix check_patterns when a class is not installed and represented with a
+  FailedImport class. By `Thomas Moreau`_ (:gh:`844`)
 
 - Fix running ``benchopt test`` when ``pytest`` is not installed
   in the conda env. By `Thomas Moreau`_ (:gh:`838`)